// src/apis/models/FireBytes_Unified_ListViews_Dtos_ListViewColumnColorDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_ListViews_Dtos_ListViewColumnColorDto = {
    id?: string | null;
    readonly extraProperties?: Record<string, any> | null;
    concurrencyStamp?: string | null;
    creationTime?: string;
    creatorId?: string | null;
    lastModificationTime?: string | null;
    lastModifierId?: string | null;
    listViewColumnsId?: string | null;
    colFieldName?: string | null;
    determineType?: string | null;
    value?: string | null;
    color?: string | null;
    tenantId?: string | null;
};