// src/apis/models/FireBytes_Unified_ListViews_Dtos_SaveListViewAsPersonalInput.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { FireBytes_Unified_ListViews_Dtos_ListViewDto } from './FireBytes_Unified_ListViews_Dtos_ListViewDto';
export type FireBytes_Unified_ListViews_Dtos_SaveListViewAsPersonalInput = {
    userId?: string | null;
    listViewDto?: FireBytes_Unified_ListViews_Dtos_ListViewDto;
};